Split some logic from VisitRepository into its own injectable repository
